Check and scan funcs
